--- plugins/device/deviceutils/utils.go
@@ -1,178 +1,0 @@
-package deviceutils
-
-import (
-	"errors"
-	"fmt"
-	"os/exec"
-	"strconv"
-	"strings"
-	"syscall"
-
-	"github.com/gluster/glusterd2/pkg/utils"
-)
-
-// constants to convert respective size into bytes.
-const (
-	KB  = 1000
-	KiB = 1024
-	MB  = 1000 * KB
-	MiB = 1024 * KiB
-	GB  = 1000 * MB
-	GiB = 1024 * MiB
-	TB  = 1000 * GB
-	TiB = 1024 * GiB
-)
-
-const (
-	maxMetadataSize = 16 * GiB
-	chunkSize       = "1280k"
-)
-
-//CreatePV is used to create physical volume.
-func CreatePV(device string) error {
-	return utils.ExecuteCommandRun("pvcreate", "--metadatasize=128M", "--dataalignment=256K", device)
-}
-
-//CreateVG is used to create volume group
-func CreateVG(device string, vgName string) error {
-	return utils.ExecuteCommandRun("vgcreate", vgName, device)
-}
-
-//RemoveVG is used to remove volume group.
-func RemoveVG(vgName string) error {
-	return utils.ExecuteCommandRun("vgremove", vgName)
-}
-
-//RemovePV is used to remove physical volume
-func RemovePV(device string) error {
-	return utils.ExecuteCommandRun("pvremove", device)
-}
-
-// GetVgAvailableSize gets available size of given Vg
-func GetVgAvailableSize(vgname string) (uint64, uint64, error) {
-	out, err := exec.Command("vgdisplay", "-c", vgname).Output()
-	if err != nil {
-		return 0, 0, err
-	}
-	vgdata := strings.Split(strings.TrimRight(string(out), "\n"), ":")
-
-	if len(vgdata) != 17 {
-		return 0, 0, errors.New("failed to get free size of VG: " + vgname)
-	}
-
-	// Physical extent size index is 12
-	extentSize, err := strconv.ParseUint(vgdata[12], 10, 64)
-	if err != nil {
-		return 0, 0, err
-	}
-
-	// Free Extents index is 15
-	freeExtents, err := strconv.ParseUint(vgdata[15], 10, 64)
-	if err != nil {
-		return 0, 0, err
-	}
-	extentSize = extentSize * KB
-
-	return extentSize * freeExtents, extentSize, nil
-}
-
-// GetPoolMetadataSize calculates the thin pool metadata size based on the given thin pool size
-func GetPoolMetadataSize(poolsize uint64) uint64 {
-	// https://access.redhat.com/documentation/en-us/red_hat_gluster_storage/3.3/html-single/administration_guide/#Brick_Configuration
-	// Minimum metadata size required is 0.5% and Max upto 16GB ~ 17179869184 Bytes
-
-	metadataSize := uint64(float64(poolsize) * 0.005)
-	rem := metadataSize % 512
-	if rem > 0 {
-		metadataSize += (512 - rem)
-	}
-
-	if metadataSize > maxMetadataSize {
-		metadataSize = maxMetadataSize
-	}
-
-	return metadataSize
-}
-
-// CreateTP creates LVM Thin Pool
-func CreateTP(vgname, tpname string, tpsize, metasize uint64) error {
-	// TODO: Chunksize adjust based on RAID/JBOD
-	return utils.ExecuteCommandRun("lvcreate",
-		"--thin", vgname+"/"+tpname,
-		"--size", fmt.Sprintf("%dB", tpsize),
-		"--poolmetadatasize", fmt.Sprintf("%dB", metasize),
-		"-c", chunkSize,
-		"--zero", "n",
-	)
-}
-
-// CreateLV creates LVM Logical Volume
-func CreateLV(vgname, tpname, lvname string, lvsize uint64) error {
-	return utils.ExecuteCommandRun("lvcreate",
-		"--virtualsize", fmt.Sprintf("%dB", lvsize),
-		"--thin",
-		"--name", lvname,
-		vgname+"/"+tpname,
-	)
-}
-
-// MakeXfs creates XFS filesystem
-func MakeXfs(dev string, mkfsOpts ...string) error {
-	mkfsOpts = append([]string{dev}, mkfsOpts...)
-	// TODO: Adjust -d su=<>,sw=<> based on RAID/JBOD
-	return utils.ExecuteCommandRun("mkfs.xfs",
-		mkfsOpts...,
-	)
-}
-
-// BrickMount mounts the brick LV
-func BrickMount(dev, mountdir string) error {
-	return utils.ExecuteCommandRun("mount",
-		"-o", "rw,inode64,noatime,nouuid",
-		dev,
-		mountdir,
-	)
-}
-
-// BrickUnmount unmounts the Brick
-func BrickUnmount(mountdir string) error {
-	return syscall.Unmount(mountdir, syscall.MNT_FORCE)
-}
-
-// RemoveLV removes Logical Volume
-func RemoveLV(vgName, lvName string) error {
-	return utils.ExecuteCommandRun("lvremove", "-f", vgName+"/"+lvName)
-}
-
-// NumberOfLvs returns number of Lvs present in thinpool
-func NumberOfLvs(vgname, tpname string) (int, error) {
-	nlv := 0
-	out, err := utils.ExecuteCommandOutput(
-		"lvs", "--no-headings", "--select",
-		fmt.Sprintf("vg_name=%s&&pool_lv=%s", vgname, tpname),
-	)
-
-	if err == nil {
-		out := strings.Trim(string(out), " \n")
-		if out == "" {
-			nlv = 0
-		} else {
-			nlv = len(strings.Split(out, "\n"))
-		}
-	}
-	return nlv, err
-}
-
-// GetThinpoolName gets thinpool name for a given LV
-func GetThinpoolName(vgname, lvname string) (string, error) {
-	out, err := utils.ExecuteCommandOutput(
-		"lvs", "--no-headings", "--select",
-		fmt.Sprintf("vg_name=%s&&lv_name=%s", vgname, lvname),
-		"-o", "pool_lv",
-	)
-	if err == nil {
-		return strings.Trim(string(out), " \n"), nil
-	}
-
-	return "", err
-}
